Moving code from interface change from previous commit
